+++ src/i-child-root-property-variant-extension.ts
